Multi-page Dash app
Part of the GUI-2.0 project.

- Modularize Dash GUI
- Add pages: Home, Side-by-Side, Info, Log
- Add experimental support for WebGL 3D

Home shows a single viewer and basic app info.
Side-by-Side is the equivalent of the old GUI.
Info shows detailed information about the app.
Log displays the complete log file.
In the future, a board and Python scripting page will be added.

The following features were removed from the GUI:
- Model saving/loading
- App framerate control
- Substeps

The following features are missing but will be added later:
- Recording
- Benchmarking / Profiling
- TensorBoard link
- 1D plotting

Bugfixes (compared to old GUI):
- When pausing, the latest frame would often not be displayed
- When reloading, all model controls were reset
- The status bar would freeze on some browsers